### 2025-08-14 92acc57
add experimental kernal for biot effective stress
