+++ include/kernels/ADsaturatedporousflow/ADHydroTimeDerivative.h
@@ -1,0 +1,17 @@
+
+
+#pragma once
+
+#include "ADTimeKernelValue.h"
+
+class ADHydroTimeDerivative : public ADTimeKernelValue
+{
+public:
+  static InputParameters validParams();
+
+  ADHydroTimeDerivative(const InputParameters & parameters);
+
+protected:
+  virtual ADReal precomputeQpResidual() override;
+  const ADMaterialProperty<Real> & _biot_modulus;
+};
