### 2025-08-17 544516b
add AD porous saturated flow kernels
