+++ src/materials/porous_flow/PorousFlowPermeabilityTensorFromVarPFF.C
@@ -1,7 +1,7 @@
 #include "PorousFlowPermeabilityTensorFromVarPFF.h"
 
 registerMooseObject("PorousFlowApp", PorousFlowPermeabilityTensorFromVarPFF);
-//registerMooseObject("PorousFlowApp", ADPorousFlowPermeabilityTensorFromVarPFF);
+registerMooseObject("PorousFlowApp", ADPorousFlowPermeabilityTensorFromVarPFF);
 
 template <bool is_ad>
 InputParameters
@@ -9,66 +9,48 @@ PorousFlowPermeabilityTensorFromVarPFFTempl<is_ad>::validParams()
 {
   InputParameters params = PorousFlowPermeabilityBase::validParams();
   params.addRequiredCoupledVar("perm", "The scalar permeability variable.");
-  params.addParam<RealTensorValue>(
-      "k_anisotropy",
-      "Anisotropy tensor to multiply the scalar permeability. "
-      "Defaults to identity tensor for isotropic permeability.");
+  params.addParam<RealTensorValue>("k_anisotropy", RealTensorValue(1,0,0, 0,1,0, 0,0,1),
+                                   "Anisotropy tensor to multiply the scalar permeability.");
+  params.addParam<MaterialPropertyName>("crack_direction_name", "direction_matrix",
+                                        "Material property name for the crack direction tensor.");
+  params.addParam<MaterialPropertyName>("coef_name", "A",
+                                        "Material property name for the coefficient.");
+  params.addParam<std::string>("model", "directional", "directional or exponential");
   params.addParam<MaterialPropertyName>(
-      "crack_direction_name", "direction_matrix",
-      "Material property name for the crack direction tensor.");
-  params.addParam<MaterialPropertyName>(
-      "coef_name", "A", "Material property name for the coefficient.");
-  params.addParam<std::string>(
-      "model", "directional",
-      "Choose the model: 'directional' or 'exponential'.");
-  params.addClassDescription(
-      "Computes permeability tensor from a coupled scalar variable and an anisotropy tensor. "
-      "An additional term can be added to account for fracture plane permeability changes.");
+    "permeability_tensor_name",
+    "permeability",
+    "Name of the permeability tensor property to declare/expose");
   return params;
 }
 
 template <bool is_ad>
-PorousFlowPermeabilityTensorFromVarPFFTempl<is_ad>::PorousFlowPermeabilityTensorFromVarPFFTempl(
-    const InputParameters & parameters)
+PorousFlowPermeabilityTensorFromVarPFFTempl<is_ad>::
+PorousFlowPermeabilityTensorFromVarPFFTempl(const InputParameters & parameters)
   : PorousFlowPermeabilityBaseTempl<is_ad>(parameters),
-    _perm(coupledValue("perm")),
+    _perm(this->template coupledGenericValue<is_ad>("perm")),
     _k_anisotropy(parameters.isParamValid("k_anisotropy")
-                      ? this->template getParam<RealTensorValue>("k_anisotropy")
-                      : RealTensorValue(1.0, 0.0, 0.0,
-                                        0.0, 1.0, 0.0,
-                                        0.0, 0.0, 1.0)),
+                    ? this->template getParam<RealTensorValue>("k_anisotropy")
+                    : RealTensorValue(1,0,0, 0,1,0, 0,0,1)),
     _crack_direction_name(this->template getParam<MaterialPropertyName>("crack_direction_name")),
[…]
