### 2025-08-17 ba283bf
add AD porous saturated flow kernels
