+++ src/kernels/ADsaturatedporousflow/ADHydroDiffusion.C
@@ -13,6 +13,11 @@ ADHydroDiffusion::validParams()
   params.addParam<MaterialPropertyName>("fluid_density", "fluid_density", "Name of fluid density as an input.");
   params.addParam<MaterialPropertyName>("fluid_viscosity", "fluid_viscosity", "Name of fluid viscosity as an input.");
   params.addParam<MaterialPropertyName>("permeability_tensor", "permeability_tensor", "Name of permeability tensor as an input.");
+  params.addParam<bool>(
+  "mass_flux",
+  false,
+  "Whether to multiply density into porousflow governing equation."
+  );
   return params;
 }
 
@@ -20,14 +25,20 @@ ADHydroDiffusion::ADHydroDiffusion(const InputParameters & parameters)
 : ADKernel(parameters), 
   _fluid_density(getADMaterialProperty<Real>(getParam<MaterialPropertyName>("fluid_density") )),
   _fluid_viscosity(getADMaterialProperty<Real>(getParam<MaterialPropertyName>("fluid_viscosity") )),
-  _permeability_tensor(getADMaterialProperty<RankTwoTensor>(getParam<MaterialPropertyName>("permeability_tensor") ))
+  _permeability_tensor(getADMaterialProperty<RankTwoTensor>(getParam<MaterialPropertyName>("permeability_tensor") )),
+  _mass_flux(getParam<bool>("mass_flux"))
 { 
 }
 
 ADReal
 ADHydroDiffusion::computeQpResidual()
-{
-  ADRankTwoTensor prop = _fluid_density[_qp]/_fluid_viscosity[_qp] * _permeability_tensor[_qp];
+{ 
+  ADRankTwoTensor prop;
+  if (_mass_flux)
+    prop = _fluid_density[_qp]/_fluid_viscosity[_qp] * _permeability_tensor[_qp];
+  else
+    prop = _permeability_tensor[_qp]/_fluid_viscosity[_qp];
+
   return _grad_test[_i][_qp] * (prop * _grad_u[_qp]);
 }
 
