### 2025-10-22 #8770 @AnjaliSandip
@DanielChou0916, if you have any suggestions for potential reviewers, please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
