### 2025-11-05 #8770 @AnjaliSandip
👋 @maruti-iitm, @shahchiragh, @shipengcheng1230, @dIOGOLOC, @fruzsinaagocs, @pmocz would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines:[ https://joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html).
