### 2025-11-05 #8770 @fruzsinaagocs
Hi Anjali, thank you for inviting me to review but at this time I have to decline, since I also edit for JOSS. I unchecked the "available to review" box on my profile -- not sure why it was set.
