### 2025-11-07 #8770 @shahchiragh
Hi Anjali, thank you for the invitation.
I can help review. What’s the tentative timeline on the review?

Thanks,
Chirag

On Wed, Nov 5, 2025 at 1:39 PM Fruzsina Agocs ***@***.***>
wrote:

> *fruzsinaagocs* left a comment (openjournals/joss-reviews#8770)
> <https://github.com/openjournals/joss-reviews/issues/8770#issuecomment-3492764139>
>
> Hi Anjali, thank you for inviting me to review but at this time I have to
> decline, since I also edit for JOSS. I unchecked the "available to review"
> box on my profile -- not sure why it was set.
>
