### 2025-11-12 #9406 @editorialbot
[REVIEW]: Felino: Extension of an open-source phase-field framework to geomaterial fracture
**Submitting author:** @DanielChou0916 (<a href="http://orcid.org/0000-0001-7358-8819">Daniel Chou</a>)
**Repository:** https://github.com/DanielChou0916/felino
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.2
