### 2025-11-12 DanielChou0916/felino#1 @pmocz
(minor) add "About" description and link to docs on the repo github page
Add "About" description, link to docs, tags for better discoverability of your repo in GitHub

[joss review thread](https://github.com/openjournals/joss-reviews/issues/9406)
