### 2025-11-12 DanielChou0916/felino#2 @pmocz
Add contributing guidelines
Add a CONTRIBUTING.md file into the repo to describe how people should contribute to your open-source project

See:
https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

[joss review thread](https://github.com/openjournals/joss-reviews/issues/9406)
