### 2025-11-12 DanielChou0916/felino#3 @pmocz
Describe state of the field in the joss paper
State of the field: Do the authors describe how this software compares to other commonly-used packages?

Add discussion / references  into your `paper.md` on to closest software in your field.

[joss review thread](https://github.com/openjournals/joss-reviews/issues/9406)
