### 2026-02-09 #9406 @AnjaliSandip
> > Hi [@AnjaliSandip](https://github.com/AnjaliSandip), I am unable tag the author on this comment. Hopefully they are able to see my comments. The repository looks healthy so far; I wasn’t able to run the `Functionality` steps on macOS due to resource constraints, but I’ll retry using Docker or a VM (Ubuntu) and report back on this soon.
> > Review feedback:
> > ```
> > * For the software paper, it appears it is missing a few things such as:
> >
> >   * Comparison to other software in the same field. I would suggest to add this to clarify strengths and tradeoffs
> >   * Missing references. Several claims and comparisons lack citations; please add in‑text references and a complete references.
> >   * The paper lists key features but doesn’t compare them to alternatives. A short “state of the field” section would address this.
> >
> > * The writing is solid. After adding the comparison and references, run a final proofreading pass for structure, flow, and minor grammar fixes.
> >
> > * The current tutorials added in the repo appears to be like cookbooks. Is this correct? If so,  I would suggest to label them as cookbooks or sample scripts, and add brief documentation for each example (purpose, inputs, expected output, and quick start).
> > ```
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> >
> > I think [@pmocz](https://github.com/pmocz) has created some issues, I would create missing ones to create pointers. Let me know if you have any questions.
>
> [@DanielChou0916](https://github.com/DanielChou0916), could you please address the above?

@DanielChou0916, I am following up on the above.  Please address the above at your earliest convenience.
