+++ paper/paper.bib
@@ -91,4 +91,99 @@
   number={11104},
   pages={111041},
   year={2025}
+
+}
+
+%ABAQUS
+@article{molnar2020open,
+  title={An open-source Abaqus implementation of the phase-field method to study the effect of plasticity on the instantaneous fracture toughness in dynamic crack propagation},
+  author={Moln{\'a}r, Gergely and Gravouil, Anthony and Seghir, Rian and R{\'e}thor{\'e}, Julien},
+  journal={Computer Methods in Applied Mechanics and Engineering},
+  volume={365},
+  pages={113004},
+  year={2020},
+  publisher={Elsevier}
+}
+
+
+@article{molnar20172d,
+  title={2D and 3D Abaqus implementation of a robust staggered phase-field solution for modeling brittle fracture},
+  author={Moln{\'a}r, Gergely and Gravouil, Anthony},
+  journal={Finite Elements in Analysis and Design},
+  volume={130},
+  pages={27--38},
+  year={2017},
+  publisher={Elsevier}
+}
+
+% Fatigue
+@article{li2023review,
+  title={A review on phase field models for fracture and fatigue},
+  author={Li, Peidong and Li, Weidong and Li, Biao and Yang, Shuo and Shen, Yongxing and Wang, Qingyuan and Zhou, Kun},
+  journal={Engineering Fracture Mechanics},
+  volume={289},
+  pages={109419},
+  year={2023},
+  publisher={Elsevier}
+}
+
+@article{kalina2023overview,
+  title={Overview of phase-field models for fatigue fracture in a unified framework},
+  author={Kalina, Martha and Schneider, Tom and Brummund, J{\"o}rg and K{\"a}stner, Markus},
+  journal={Engineering Fracture Mechanics},
+  volume={288},
+  pages={109318},
+  year={2023},
+  publisher={Elsevier}
+}
+
+% ABAQUS
+@book{ABSstd2009,
+  title = "ABAQUS/Standard User's Manual, Version 6.9",
+  author = "Michael Smith",
+  year = "2009",
+  language = "English",
+  publisher = "Dassault Syst{\`e}mes Simulia Corp",
+  address = "United States",
+}
+
+%COMSOL
+@software{COMSOL,
+author = {COMSOL AB, Stockholm, Sweden},
+title = {COMSOL Multiphysics® software},
+version = {6.4},
+year = {2025},
+url = {www.comsol.com}
+}
+
+% FEniCS
+
+@misc{BarattaEtal2023,
+  title     = {{DOLFINx}: the next generation {FEniCS} problem solving environment},
+  author    = {Baratta, Igor A. and Dean, Joseph P. and Dokken, J{\o}rgen S. and Habera, Michal and Hale, Jack S. and Richardson, Chris N. and Rognes, Marie E. and Scroggs, Matthew W. and Sime, Nathan and Wells, Garth N.},
+  doi       = {10.5281/zenodo.10447666},
+  year      = {2023},
+  howpublished = {preprint}
+}
+
+@article{ScroggsEtal2022,
+  title     = {Construction of arbitrary order finite element degree-of-freedom maps on polygonal and polyhedral cell meshes},
+  author    = {Scroggs, Matthew W. and Dokken, J{\o}rgen S. and Richardson, Chris N. and Wells, Garth N.},
+  journal   = {ACM Transactions on Mathematical Software},
+  year      = {2022},
+  volume    = {48},
+  number    = {2},
+  doi       = {10.1145/3524456},
+  pages     = {{18:1--18:23}},
+}
+
+@article{BasixJoss,
+  title     = {Basix: a runtime finite element basis evaluation library},
[…]
