### 2026-02-11 #9406 @AnjaliSandip
@shahchiragh @dIOGOLOC, is the paper/software ready for publication?
