### 2026-03-09 #9406 @AnjaliSandip
> [@dIOGOLOC](https://github.com/dIOGOLOC), is the paper/software ready for publication?

@dIOGOLOC, I am following up on this.  Is the paper/software ready for publication?
