### 2026-03-20 #9406 @dIOGOLOC
- [X] I confirm that I read and will adhere to the JOSS code of conduct.

- [X] **Repository:** Is the source code for this software available at the https://github.com/DanielChou0916/felino?

- [X] **Community guidelines:** Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support.
