### 2026-03-20 #9406 @dIOGOLOC
> [@dIOGOLOC](https://github.com/dIOGOLOC), is the paper/software ready for publication?

All review comments have been satisfactorily addressed.
