+++ paper/paper.bib
@@ -6,6 +6,7 @@
   volume={11},
   pages={100430},
   year={2020},
+  doi       = {10.1016/j.softx.2020.100430},
   publisher={Elsevier}
 }
 
@@ -16,6 +17,7 @@
   volume={22},
   pages={237--254},
   year={2006},
+  doi = {10.1007/s00366-006-0049-3},
   publisher={Springer}
 }
 
@@ -25,6 +27,7 @@
   title={Felino: A Modular Open-Access Software for High-Cycle Fatigue Simulation Via Phase-Field Methods},
   author={Chou, Daniel t and Darabi, Roya},
   journal={Available at SSRN 5291375},
+  doi = {10.2139/ssrn.5281200},
   year={2025}
 }
 
@@ -37,6 +40,7 @@
   number={1},
   pages={83--100},
   year={2022},
+  doi = {10.1007/s10704-022-00653-z},
   publisher={Springer}
 }
 
@@ -57,6 +61,7 @@
   volume={309},
   pages={113170},
   year={2025},
+  doi = {10.1016/j.ijsolstr.2024.113170},
   publisher={Elsevier}
 }
 
@@ -68,6 +73,7 @@
   volume={121},
   pages={103555},
   year={2022},
+  doi = {10.1016/j.tafmec.2022.103555},
   publisher={Elsevier}
 }
 
@@ -80,6 +86,7 @@
   number={5},
   pages={779--805},
   year={2020},
+  doi = {10.1002/nme.6244},
   publisher={Wiley Online Library}
 }
 
@@ -90,6 +97,7 @@
   volume={320},
   number={11104},
   pages={111041},
+  doi = {10.2139/ssrn.4980916},
   year={2025}
 
 }
@@ -102,6 +110,7 @@
   volume={365},
   pages={113004},
   year={2020},
+  doi = {10.1016/j.cma.2020.113004},
   publisher={Elsevier}
 }
 
@@ -113,6 +122,7 @@
   volume={130},
   pages={27--38},
   year={2017},
+  doi = {10.1016/j.finel.2017.03.002},
   publisher={Elsevier}
 }
 
@@ -134,6 +144,7 @@
   volume={288},
   pages={109318},
   year={2023},
+  doi = {10.1016/j.engfracmech.2023.109318},
   publisher={Elsevier}
 }
 
